Move rendering from AP to AV
