+++ actionpack/lib/action_controller/metal/rendering.rb
@@ -2,7 +2,7 @@ module ActionController
   module Rendering
     extend ActiveSupport::Concern
 
-    include AbstractController::Rendering
+    include ActionView::Rendering
 
     # Before processing, set the request formats in current controller formats.
     def process_action(*) #:nodoc:
